Add confirm userID in DeleteAccountTryWindow.

+++ src/components/popups/DeleteAccountTryWindow.tsx
@@ -6,13 +6,19 @@ import { PopupTemplate } from "./PopupTemplate";
 
 export const DeleteAccountTryWindow = (props: {
   setPopupStatus: React.Dispatch<React.SetStateAction<PopupStatus>>;
+  realUserID: string;
 }) => {
   const { deleteAccount, status } = useDeleteAccount();
   const [loadingStatus, setLoadingStatus] = useState<
     "notYet" | "loading" | "finish"
   >("notYet");
+  const [confirmUserID, setConfirmUserID] = useState("");
 
   const handleClickDeleteAccount = () => {
+    if (confirmUserID !== props.realUserID) {
+      console.log("User ID wrong.");
+      return;
+    }
     setLoadingStatus("loading");
     deleteAccount().then(({ status }) => {
       setLoadingStatus("finish");
@@ -28,23 +34,37 @@ export const DeleteAccountTryWindow = (props: {
 
   return (
     <PopupTemplate handleClose={() => props.setPopupStatus(null)}>
-      <div className="text-center">Delete Account</div>
+      <div className="">
+        <div className="text-center">Delete Account</div>
+        <div className="flex flex-col">
+          <div>
+            This is very dangerous operation. You can NOT recover your account.
+            If you'd like to delete this account, please fill in the below blank
+            with your user ID.
+          </div>
+          <input
+            className="border-2 border-gray-300 rounded-full px-3 focus:outline-none focus:border-rose-600"
+            value={confirmUserID}
+            onChange={(e) => setConfirmUserID(e.currentTarget.value)}
+          />
+        </div>
 
-      <div className="flex items-center justify-around h-20">
-        <button
-          className="border-2 border-gray-400 rounded-md px-2 mx-4 hover:bg-rose-200"
-          onClick={() => {
-            handleClickDeleteAccount();
-          }}
-        >
-          Delete
-        </button>
-        <button
-          className="border-2 border-gray-400 rounded-md px-2 mx-4 hover:bg-rose-200"
-          onClick={() => props.setPopupStatus(null)}
-        >
-          Cancel
-        </button>
+        <div className="flex items-center justify-around h-20">
+          <button
+            className="border-2 border-gray-400 rounded-md px-2 mx-4 hover:bg-rose-200"
+            onClick={() => {
+              handleClickDeleteAccount();
+            }}
+          >
+            Delete
+          </button>
+          <button
+            className="border-2 border-gray-400 rounded-md px-2 mx-4 hover:bg-rose-200"
+            onClick={() => props.setPopupStatus(null)}
+          >
+            Cancel
+          </button>
+        </div>
       </div>
     </PopupTemplate>
   );
